backend/requirements.txt
Flask==3.0.0
psutil
flask_sqlalchemy
Flask-CORS>=4.0.1
Flask-SocketIO==5.3.5
python-socketio==5.10.0
opencv-python==4.10.0.84
deepface==0.0.93   
tensorflow==2.20.0
numpy
selenium
Pillow>=10.4.0
python-dotenv==1.0.0
SQLAlchemy>=2.0.36
pandas>=2.2.0
PyJWT==2.8.0
tf-keras==2.20.0
requests==2.32.3
mss>=6.1.0
colorlog>=6.6.0
pywinauto>=0.6.8
urllib3>=1.26.0